DefaultKeyBinding.dict: Change bindings for new line before & after ¶
Apple-Return is a fairly common binding and I don't want it to conflict
with anything. I use it quite often when making comments on GitHub and
today noticed a little hiccup when trying to submit a comment. When I
pressed Apple-Return a second time the comment submitted successfully,
but I'd prefer not to take my chances.

That said, these seem like they could be useful bindings to have, so
bind them to something close to the existing 'o' & 'O' bindings.