DefaultKeyBinding.dict: Change delete word forward to Opt-d
Move the `deleteWordForward:` command to Option-d and change the
existing command for that (delete paragraph) to Control-Option-d.

I wanted to match the Readline/Emacs keys that I'm used to for deleting
a word forward. Thus this becomes Meta-d to delete a word forward which
is just like the Emacs style and requires no extra memorisation. On top
of that, it's easier to type because it requires one less modifier key.

Since I feel like I'll be using this one more than 'delete paragraph',
the switch made sense. Also, since deleting a whole paragraph is a more
destructive change, adding an extra modifier key in order to accomplish
it doesn't seem like a bad thing to me.